check whether assemblyinfo existed before modifying it

+++ tools/BuildScripts/AdaptAutorestModule.ps1
@@ -129,7 +129,10 @@ New-ModuleManifest -Path $parentModulePsd1Path @parentModuleMetadata
 # Update module version in submodule AssemblyInfo.cs with parent module version
 $moduleVersion = $parentModuleMetadata.ModuleVersion
 $assemblyInfoPath = Join-Path $subModulePath 'Properties' 'AssemblyInfo.cs'
-(Get-Content $assemblyInfoPath -Raw) -replace "0.1.0.0", $moduleVersion | Set-Content $assemblyInfoPath -Force
+if (Test-Path $assemblyInfoPath) {
+    (Get-Content $assemblyInfoPath -Raw) -replace "0.1.0.0", $moduleVersion | Set-Content $assemblyInfoPath -Force
+}
+
 
 <#
     merge temporary sub module csproj to parent module sln (for platyPS help markdown generation)
